design: start page
start page 모듈화 진행.
Heading 화면 가운데로 오게 조정.
