+++ src/pages/Start/Main/Heading/index.jsx
@@ -1,0 +1,3 @@
+import Heading from './Heading';
+
+export default Heading;
