+++ src/pages/Start/Main/Main.scss
@@ -1,0 +1,71 @@
+main.Main {
+  /* Properties */
+  //grid-item
+  grid-column: 2 / 3;
+  grid-row: 1 / 3;
+  //flex-container
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  /* Selector */
+  > div {
+    > .config {
+      /* Property */
+      //box
+      height: 500px;
+      width: 100%;
+      margin: 50px 0;
+      //text
+      text-align: center;
+      //animation
+      transition: all 2s ease-in-out;
+
+      /* Selector */
+      &.onPress {
+        height: 0px;
+        opacity: 0;
+      }
+    }
+    > .button {
+      /* Property */
+      //loading effects
+      animation: 8s ease-in-out loadEffect1;
+      @keyframes loadEffect1 {
+        0% {
+          opacity: 0;
+        }
+        50% {
+          opacity: 0;
+        }
+        100% {
+          opacity: 1;
+        }
+      }
+      //box
+      margin: 50px 0;
+      //animation
+      transition: all 3s ease-out;
+      opacity: 1;
+      //flex
+      display: flex;
+      justify-content: center;
+      align-items: center;
+
+      /* Selector */
+      &.off {
+        pointer-events: none;
+        opacity: 0;
+      }
+    }
+    > .test {
+      width: 600px;
+      height: 200px;
+      border-radius: 12px;
+      padding: 30px;
+      font-family: Nanum Gothic Coding;
+      line-height: 1.5;
+      background-color: darken(#202020, 7%);
+    }
+  }
+}
