+++ src/pages/Start/Start.scss
@@ -1,5 +1,3 @@
-@import '@/styles/Mixin';
-
 .Start {
   /* Variables */
   $width: 80px;
@@ -13,122 +11,4 @@
   display: grid;
   grid-template-columns: $width 1fr $width;
   grid-template-rows: $height 1fr $height;
-
-  /* Selector */
-  > .main {
-    /* Properties */
-    //grid-item
-    grid-column: 2 / 3;
-    grid-row: 2 / 3;
-    //flex-container
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    /* Selector */
-    > div {
-      > .heading {
-        /* Property */
-        //box
-        margin: 50px 0;
-        //others
-        pointer-events: none;
-
-        /* Selector */
-        > .mock {
-          //loading effects
-          animation: 2s ease-in-out loadEffect1;
-          @keyframes loadEffect1 {
-            0% {
-              opacity: 0;
-            }
-            100% {
-              opacity: 1;
-            }
-          }
-          //box
-          margin-bottom: 10px;
-          //others
-          transform: rotate(-10deg);
-        }
-        > .interview {
-          //loading effects
-          animation: 4s ease-in-out loadEffect1;
-          @keyframes loadEffect1 {
-            0% {
-              opacity: 0;
-            }
-            50% {
-              opacity: 0;
-            }
-            100% {
-              opacity: 1;
-            }
-          }
-          //box
-          margin-left: 50px;
-        }
-      }
-      > .config {
-        /* Include */
-        @include type2BackgroundColor;
-
-        /* Property */
-        //box
-        height: 100px;
-        width: 100%;
-        margin: 50px 0;
-        //text
-        text-align: center;
-        //animation
-        transition: all 2s ease-in-out;
-
-        /* Selector */
-        &.onPress {
-          height: 0px;
-          opacity: 0;
-        }
-      }
-      > .button {
-        /* Property */
-        //loading effects
-        animation: 8s ease-in-out loadEffect1;
-        @keyframes loadEffect1 {
-          0% {
-            opacity: 0;
-          }
-          50% {
-            opacity: 0;
-          }
-          100% {
-            opacity: 1;
-          }
-        }
-        //box
-        margin: 50px 0;
-        //animation
-        transition: all 3s ease-out;
-        opacity: 1;
-        //flex
-        display: flex;
-        justify-content: center;
-        align-items: center;
-
-        /* Selector */
-        &.off {
-          pointer-events: none;
-          opacity: 0;
-        }
-      }
-      > .test {
-        width: 600px;
-        height: 200px;
-        border-radius: 12px;
-        padding: 30px;
-        font-family: Nanum Gothic Coding;
-        line-height: 1.5;
-        background-color: darken(#202020, 7%);
-      }
-    }
-  }
 }
